initial commit
Everything works correctly so far. Only the basics though. There's a light at the viewer (+z-axis) and the object gets loaded in the center. You can move with the keyboard ('wasd' and 'qe' do translation, 'ijkl' and 'ou' do rotations). Planning on changing that to use mouse input. Also hoping to add a way to move the lighting around.